Isoate merge for refactor
