+++ command_router.js
@@ -8,7 +8,7 @@ const test = require("./commands/test");
 const createprogram = require("./commands/createprogram");
 const deleteprogram = require("./commands/deleteprogram");
 const programsettings = require("./commands/programsettings.js");
-const merge = require("./commands/merge");
+const merge = require("./commands/merge/merge.js");
 const contrib = require("./commands/contrib");
 
 const CommandError = require("./command_error.js");
